+++ lib/blade/folder/plugin.ts
@@ -1,5 +1,5 @@
 import {ValueMap} from '../../common/model/value-map';
-import {ParamsParsers, parseParams} from '../../common/params';
+import {ParamsParsers, parseParams} from '../../common/params-parsers';
 import {BladeParams} from '../common/api/params';
 import {BladePlugin} from '../plugin';
 import {FolderApi} from './api/folder';
