+++ lib/blade/separator/plugin.ts
@@ -1,4 +1,4 @@
-import {ParamsParsers, parseParams} from '../../common/params';
+import {ParamsParsers, parseParams} from '../../common/params-parsers';
 import {BladeParams} from '../common/api/params';
 import {BladePlugin} from '../plugin';
 import {SeparatorApi} from './api/separator';
