+++ lib/monitor-binding/boolean/plugin.ts
@@ -3,7 +3,7 @@ import {
 	BooleanFormatter,
 	boolFromUnknown,
 } from '../../common/converter/boolean';
-import {ParamsParsers, parseParams} from '../../common/params';
+import {ParamsParsers, parseParams} from '../../common/params-parsers';
 import {Constants} from '../../misc/constants';
 import {MultiLogController} from '../common/controller/multi-log';
 import {SingleLogMonitorController} from '../common/controller/single-log';
